# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: DamDam API
  version: 0.1.0
paths:
...
  /v1/sos:
    post:
      tags:
      - SOS
      summary: Create Sos
      operationId: create_sos_v1_sos_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SOSCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SOSCreateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /v1/sos/{alert_id}/cancel:
    post:
      tags:
      - SOS
      summary: Cancel Sos
      operationId: cancel_sos_v1_sos__alert_id__cancel_post
      security:
      - HTTPBearer: []
      parameters:
      - name: alert_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Alert Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SOSStatusResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/hto/sos-alerts:
    get:
      tags:
      - SOS
      summary: List Sos
      operationId: list_sos_v1_hto_sos_alerts_get
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/SOSStatus'
          - type: 'null'
          default: active
          title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HtoSOSAlertsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/hto/sos-alerts/{alert_id}/resolve:
    post:
      tags:
      - SOS
      summary: Resolve Sos
      operationId: resolve_sos_v1_hto_sos_alerts__alert_id__resolve_post
      security:
      - HTTPBearer: []
      parameters:
      - name: alert_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Alert Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SOSStatusResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/hto/push-subscriptions:
    post:
      tags:
      - SOS
      summary: Create Push Subscription
      operationId: create_push_subscription_v1_hto_push_subscriptions_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PushSubscriptionCreate'
        required: true
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    HtoSOSAlertResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        pilgrim_name:
          type: string
          title: Pilgrim Name
        pilgrim_phone:
          type: string
          title: Pilgrim Phone
        timestamp:
          type: string
          format: date-time
          title: Timestamp
        latitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Latitude
        longitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Longitude
        status:
          $ref: '#/components/schemas/SOSStatus'
      type: object
      required:
      - id
      - pilgrim_name
      - pilgrim_phone
      - timestamp
      - latitude
      - longitude
      - status
      title: HtoSOSAlertResponse
    HtoSOSAlertsResponse:
      properties:
        alerts:
          items:
            $ref: '#/components/schemas/HtoSOSAlertResponse'
          type: array
          title: Alerts
      type: object
      required:
      - alerts
      title: HtoSOSAlertsResponse
...
    PushSubscriptionCreate:
      properties:
        fcm_token:
          type: string
          minLength: 1
          title: Fcm Token
      type: object
      required:
      - fcm_token
      title: PushSubscriptionCreate
...
    SOSCreate:
      properties:
        client_generated_id:
          type: string
          format: uuid
          title: Client Generated Id
        timestamp:
          type: string
          format: date-time
          title: Timestamp
        latitude:
          anyOf:
          - type: number
            maximum: 90.0
            minimum: -90.0
          - type: 'null'
          title: Latitude
        longitude:
          anyOf:
          - type: number
            maximum: 180.0
            minimum: -180.0
          - type: 'null'
          title: Longitude
      type: object
      required:
      - client_generated_id
      - timestamp
      title: SOSCreate
    SOSCreateResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        status:
          $ref: '#/components/schemas/SOSStatus'
      type: object
      required:
      - id
      - status
      title: SOSCreateResponse
...
    SOSStatus:
      type: string
      enum:
      - active
      - resolved
      - cancelled
      title: SOSStatus
    SOSStatusResponse:
      properties:
        status:
          $ref: '#/components/schemas/SOSStatus'
      type: object
      required:
      - status
      title: SOSStatusResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer